ci: simplify scripts/publish.js
