--- scripts/types.d.ts
@@ -1,9 +1,0 @@
-export type Package = {
-  name: string
-  packageDir: string
-}
-
-export type BranchConfig = {
-  prerelease: boolean
-  previousVersion?: boolean
-}
